clean up kernel tools
